+++ kdl-script/examples/simple.kdl
@@ -1,4 +1,3 @@
-@derive "Display"
 struct "Point" {
     x "f64"
     y "f64"
@@ -15,7 +14,7 @@ fn "main" {
         x 10.0
         y 20.0
     }
-    
+
     let "sum" "add:" "pt1" "pt2"
     print "sum"
